Pair review DG DR

+++ src/pythonGui/karabogui/dialogs/update_dialog.py
@@ -1,17 +1,19 @@
-import os
-from os import path as op
-
 from argparse import ArgumentParser
 from contextlib import contextmanager
-from karabogui import icons
-from lxml import etree
+import os
+import os.path as op
 import pkg_resources
 import re
 import requests
+from subprocess import check_output, STDOUT, CalledProcessError
+
+from lxml import etree
 from PyQt4 import uic
 from PyQt4.QtCore import QProcess, pyqtSlot
 from PyQt4.QtGui import QDialog
-from subprocess import check_output, STDOUT, CalledProcessError
+
+from karabogui import icons
+
 
 _TIMEOUT = 0.1
 _TAG_REGEX = '^\d+\.\d+\.\d+$'
@@ -65,12 +67,11 @@ def get_latest_version():
 
 
 def _download_file_for_tag(tag):
-    """Downloads the wheel for the given tag and writes it to a file,
-    removing it when the context is finished."""
+    """Downloads the wheel for the given tag and writes it to a file
+
+    The file is removed on context exit."""
     wheel_file = _WHEEL_TEMPLATE.format(tag)
-    wheel_path = '{}{}/{}'.format(_REMOTE_SVR,
-                                  tag,
-                                  wheel_file)
+    wheel_path = '{}{}/{}'.format(_REMOTE_SVR, tag, wheel_file)
 
     try:
         wheel_request = requests.get(wheel_path, stream=True, timeout=_TIMEOUT)
@@ -98,12 +99,14 @@ def download_file_for_tag(tag):
         os.remove(wheel_file)
 
 
-def update_package(tag: str):
+def update_package(tag):
     """Updates the package to the given tag version.
 
     :returns str:
         The output given by the process.
     """
+    assert isinstance(tag, str)
+
     with download_file_for_tag(tag) as (wheel_file, err):
         if err is not None:
             return err
@@ -118,18 +121,19 @@ def update_package(tag: str):
 
 
 class UpdateDialog(QDialog):
-    """
+    """Update dialog for the Karabo GUI extensions package
+
     This dialog is used for checking the current and latest versions of
     karabo-extensions package. The user may also update the package, where a
     pip install install call is done.
     """
+
     def __init__(self, parent=None):
         super(UpdateDialog, self).__init__(parent)
         uic.loadUi(op.join(op.dirname(__file__), 'update_dialog.ui'), self)
 
         self.lb_current.setText(UNDEFINED_VERSION)
         self.lb_latest.setText(UNDEFINED_VERSION)
-
         self.bt_refresh.setIcon(icons.refresh)
         self.bt_clear_log.setIcon(icons.editClear)
 
@@ -138,7 +142,6 @@ class UpdateDialog(QDialog):
 
         # Connect signals
         self.bt_close.clicked.connect(self.accept)
-        self.bt_refresh.clicked.connect(self.refresh_versions)
         self.bt_clear_log.clicked.connect(self.ed_log.clear)
 
         # Store the current running process
@@ -147,18 +150,10 @@ class UpdateDialog(QDialog):
 
         self.refresh_versions()
 
-    @pyqtSlot()
-    def on_bt_refresh_clicked(self):
-        self.refresh_versions(should_log=True)
-
-    def refresh_versions(self, should_log=False):
-        """Method responsible for refreshing the current and latest package
-        versions"""
+    def refresh_versions(self):
+        """Refreshing the current and latest package versions"""
         self._update_current_version()
         self._update_latest_version()
-
-        if should_log:
-            self.ed_log.append('Versions refreshed')
 
         current = self.lb_current.text()
         latest = self.lb_latest.text()
@@ -171,6 +166,7 @@ class UpdateDialog(QDialog):
         self.lb_current.setVisible(False)
 
         current_version = get_current_version()
+        self._update_log('Versions refreshed')
 
         self.lb_current.setVisible(True)
         self.lb_current.setText(current_version)
@@ -193,44 +189,50 @@ class UpdateDialog(QDialog):
             self.ed_log.append(err)
             return None
 
-        # Create a process and connect its signals
-        self._process = QProcess(self)
-        self._process.setProcessChannelMode(QProcess.MergedChannels)
-        self._process.readyReadStandardOutput.connect(self._on_output)
-        self._process.finished.connect(self._on_finished)
-        self._process.error.connect(self._on_finished)
+        if self._process is None:
+            # Create a process and connect its signals
+            self._process = QProcess(self)
+            self._process.setProcessChannelMode(QProcess.MergedChannels)
+            self._process.readyRead.connect(self._on_output)
+            self._process.finished.connect(self._on_finished)
 
         self._process.start('pip install --upgrade {}'
                             .format(self._wheel_file))
 
+    def _update_log(self, text):
+        self.ed_log.append(text)
+
+    # --------------------------------------------------------------------
+    # Qt Slots
+
+    @pyqtSlot()
+    def on_bt_refresh_clicked(self):
+        self.refresh_versions()
+
     @pyqtSlot()
     def on_bt_stop_clicked(self):
         """Kills the running process"""
-        if self._process is not None:
-            if self._process.state() == QProcess.Running:
-                self._process.kill()
-                self._on_finished()
+        if (self._process is not None and
+                    self._process.state() == QProcess.Running):
+            self._process.kill()
+            self._on_finished()
 
     @pyqtSlot()
     def on_bt_update_clicked(self):
-        """Updates guiextensions to the latest tag"""
+        """Updates gui extensions to the latest tag"""
         self.bt_refresh.setEnabled(False)
         self.bt_update.setEnabled(False)
         self.bt_stop.setEnabled(True)
-
         self._start_update_process()
 
+    @pyqtSlot()
     def _on_output(self):
-        """Called whenever there is output available to be consumed from the
-        process stdout. This output is then logged in the text edit."""
-        if self._process and self._process.bytesAvailable():
-            output = self._process.readAllStandardOutput()
-            output = output.data().decode()
+        size = self._process.size()
+        data = self._process.readData(size)
+        self._update_log(data.decode())
 
-            if output:
-                self.ed_log.append(output)
-
-    def _on_finished(self, *args, **kwargs):
+    @pyqtSlot()
+    def _on_finished(self):
         """Method called whenever the process finishes or crashes. It's used
         to clean any created resources and restore the interface states."""
         if self._wheel_file and os.path.isfile(self._wheel_file):
@@ -241,15 +243,14 @@ class UpdateDialog(QDialog):
 
         self.bt_refresh.setEnabled(True)
         self.bt_stop.setEnabled(False)
-
         self.refresh_versions()
 
 
 def main():
     description = """Command-line tool to update the GUI-Extensions package"""
     ap = ArgumentParser(description=description)
-    tag_help = 'Desired package version (tag)'
-    ap.add_argument('-t', '--tag', nargs=1, required=True, help=tag_help)
+    ap.add_argument('-t', '--tag', nargs=1, required=True, type=str,
+                    help='Desired package version (tag)')
 
     args = ap.parse_args()
